Move Effect &c. into Control.Algebra.Internal.
